library_management_system/users/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-phonenumber-field==5.0.0
pkg-resources==0.0.0
pytz==2021.1
sqlparse==0.4.1
